+++ src/main/java/com/tiendis/tiendis/DAO/CarritoDetalleDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.CarritoDetalle;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarritoDetalleDAO extends CrudRepository<CarritoDetalle, Long> {
+public interface CarritoDetalleDAO extends CrudRepository<CarritoDetalle, Integer> {
 }
