+++ src/main/java/com/tiendis/tiendis/DAO/EstadoCarritoDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.EstadoCarrito;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EstadoCarritoDAO extends CrudRepository<EstadoCarrito, Long> {
+public interface EstadoCarritoDAO extends CrudRepository<EstadoCarrito, Integer> {
 }
