+++ src/main/java/com/tiendis/tiendis/DAO/ProductoDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.Producto;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductoDAO extends CrudRepository<Producto, Long> {
+public interface ProductoDAO extends CrudRepository<Producto, Integer> {
 }
