+++ src/main/java/com/tiendis/tiendis/DAO/ProveedorDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.Proveedor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProveedorDAO extends CrudRepository<Proveedor, Long> {
+public interface ProveedorDAO extends CrudRepository<Proveedor, Integer> {
 }
