+++ src/main/java/com/tiendis/tiendis/DAO/TipoCuentaDAO.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.DAO;
 import com.tiendis.tiendis.entity.TipoCuenta;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TipoCuentaDAO extends CrudRepository<TipoCuenta, Long> {
+public interface TipoCuentaDAO extends CrudRepository<TipoCuenta, Integer> {
 }
