+++ src/main/java/com/tiendis/tiendis/service/BancoService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.Banco;
 
-public interface BancoService extends GenericServiceApi <Banco, Long> {
+public interface BancoService extends GenericServiceApi <Banco, Integer> {
 }
