+++ src/main/java/com/tiendis/tiendis/service/CarritoDetalleService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.CarritoDetalle;
 
-public interface CarritoDetalleService extends GenericServiceApi<CarritoDetalle, Long> {
+public interface CarritoDetalleService extends GenericServiceApi<CarritoDetalle, Integer> {
 }
