+++ src/main/java/com/tiendis/tiendis/service/ClienteService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.Cliente;
 
-public interface ClienteService extends GenericServiceApi<Cliente, Long> {
+public interface ClienteService extends GenericServiceApi<Cliente, Integer> {
 }
