+++ src/main/java/com/tiendis/tiendis/service/FormaPagoService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.FormaPago;
 
-public interface FormaPagoService extends GenericServiceApi<FormaPago, Long> {
+public interface FormaPagoService extends GenericServiceApi<FormaPago, Integer> {
 }
