+++ src/main/java/com/tiendis/tiendis/service/Implement/BancoServiceImpl.java
@@ -9,13 +9,13 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class BancoServiceImpl extends GenericServiceImpl<Banco, Long> implements BancoService {
+public class BancoServiceImpl extends GenericServiceImpl<Banco, Integer> implements BancoService {
 
     @Autowired
     private BancoDAO bancoDAO;
 
     @Override
-    public CrudRepository<Banco, Long> getDao() {
+    public CrudRepository<Banco, Integer> getDao() {
         return bancoDAO;
     }
 }
