+++ src/main/java/com/tiendis/tiendis/service/SubCategoriaService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.SubCategoria;
 
-public interface SubCategoriaService extends GenericServiceApi<SubCategoria, Long> {
+public interface SubCategoriaService extends GenericServiceApi<SubCategoria, Integer> {
 }
