+++ src/main/java/com/tiendis/tiendis/service/TipoClienteService.java
@@ -3,5 +3,5 @@ package com.tiendis.tiendis.service;
 import com.tiendis.tiendis.commons.GenericServiceApi;
 import com.tiendis.tiendis.entity.TipoCliente;
 
-public interface TipoClienteService extends GenericServiceApi <TipoCliente, Long> {
+public interface TipoClienteService extends GenericServiceApi <TipoCliente, Integer> {
 }
